MISC: Refactors phpunit tests
